# backend/src/main/java/com/carlos/Chat/Models/chat/StringResponse.java
package com.carlos.Chat.Models.chat;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class StringResponse {
    private String response;
}