# backend/src/main/java/com/carlos/Chat/Models/user/UserResponse.java
package com.carlos.Chat.Models.user;

import java.time.LocalDateTime;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class UserResponse {
    
    private String id;
    private String name;
    private String lastName;
    private String email;
    private LocalDateTime lastSeen;
    private boolean isOnline;
}