# frontend/src/openApi/openApi.json
{
  "openapi": "3.0.1",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [{ "url": "http://localhost:8080", "description": "Generated server url" }],
  "paths": {
    "/api/v1/messages": {
      "post": {
        "tags": ["Message"],
        "operationId": "saveMessage",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/MessageRequest" } }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      },
      "patch": {
        "tags": ["Message"],
        "operationId": "setMessagesToSeen",
        "parameters": [
          { "name": "chatId", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/messages/upload-media": {
      "post": {
        "tags": ["Message"],
        "operationId": "uploadMedia",
        "parameters": [
          { "name": "chatId", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": ["file"],
                "type": "object",
                "properties": { "file": { "type": "string", "format": "binary" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/v1/chats": {
      "get": {
        "tags": ["Chat"],
        "operationId": "getChatsByReceiver",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ChatResponse" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Chat"],
        "operationId": "create",
        "parameters": [
          { "name": "senderId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "receiverId", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/StringResponse" } }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": { "type": "string" } }
              }
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "tags": ["User"],
        "operationId": "getAllUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UserResponse" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/chat/{chatId}": {
      "get": {
        "tags": ["Message"],
        "operationId": "getMessages",
        "parameters": [
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/MessageResponse" }
                }
              }
            }
          }
        }
      }
    },
    "/api/user": {
      "get": {
        "tags": ["Authentication"],
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": { "type": "object" } }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "get": {
        "tags": ["Authentication"],
        "operationId": "login",
        "responses": { "200": { "description": "OK" } }
      }
    }
  },
  "components": {
    "schemas": {
      "MessageRequest": {
        "type": "object",
        "properties": {
          "content": { "type": "string" },
          "senderId": { "type": "string" },
          "receiverId": { "type": "string" },
          "type": { "type": "string", "enum": ["TEXT", "IMAGE", "AUDIO", "VIDEO"] },
          "chatId": { "type": "string" }
        }
      },
      "StringResponse": { "type": "object", "properties": { "response": { "type": "string" } } },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string" },
          "lastSeen": { "type": "string", "format": "date-time" },
          "online": { "type": "boolean" }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "content": { "type": "string" },
          "type": { "type": "string", "enum": ["TEXT", "IMAGE", "AUDIO", "VIDEO"] },
          "state": { "type": "string", "enum": ["SENT", "SEEN"] },
          "senderId": { "type": "string" },
          "receiverId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "media": { "type": "array", "items": { "type": "string", "format": "byte" } }
        }
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "unreadCount": { "type": "integer", "format": "int64" },
          "lastMessage": { "type": "string" },
          "lastMessageTime": { "type": "string", "format": "date-time" },
          "senderId": { "type": "string" },
          "receiverId": { "type": "string" },
          "recipientOnline": { "type": "boolean" }
        }
      }
    }
  }
}